arrange files location
