+++ front/react_front/src/Components/Page/UserMyPage.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { useHistory } from "react-router-dom";
-import TrainingRegistration from "./TrainingRegistration";
-import CalendarComponent from "./CalendarComponent";
+import TrainingRegistration from "../Component/TrainingRegistration";
+import CalendarComponent from "../Component/CalendarComponent";
 import { makeStyles } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
